feat: remove redundant AstGrep method
