+++ crates/core/src/replacer/template.rs
@@ -194,10 +194,7 @@ if (true) {
 
   fn test_str_replace(replacer: &str, vars: &[(&str, &str)], expected: &str) {
     let mut env = MetaVarEnv::new();
-    let roots: Vec<_> = vars
-      .iter()
-      .map(|(v, p)| (v, Tsx.ast_grep(p).inner))
-      .collect();
+    let roots: Vec<_> = vars.iter().map(|(v, p)| (v, Tsx.ast_grep(p))).collect();
     for (var, root) in &roots {
       env.insert(var, root.root());
     }
@@ -256,10 +253,7 @@ if (true) {
 
   fn test_ellipsis_replace(replacer: &str, vars: &[(&str, &str)], expected: &str) {
     let mut env = MetaVarEnv::new();
-    let roots: Vec<_> = vars
-      .iter()
-      .map(|(v, p)| (v, Tsx.ast_grep(p).inner))
-      .collect();
+    let roots: Vec<_> = vars.iter().map(|(v, p)| (v, Tsx.ast_grep(p))).collect();
     for (var, root) in &roots {
       env.insert_multi(var, root.root().children().collect());
     }
@@ -305,10 +299,7 @@ if (true) {
 
   fn test_template_replace(template: &str, vars: &[(&str, &str)], expected: &str) {
     let mut env = MetaVarEnv::new();
-    let roots: Vec<_> = vars
-      .iter()
-      .map(|(v, p)| (v, Tsx.ast_grep(p).inner))
-      .collect();
+    let roots: Vec<_> = vars.iter().map(|(v, p)| (v, Tsx.ast_grep(p))).collect();
     for (var, root) in &roots {
       env.insert(var, root.root());
     }
